pyproject.toml
[project]
name = "denbot-rl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "debugpy>=1.8.13",
  "gputil>=1.4.0",
  "hydra-core>=1.3.2",
  "ipython>=9.0.2",
  "matplotlib>=3.10.1",
  "pyqt5>=5.15.11",
  "pyright>=1.1.397",
  "pytest>=8.3.5",
  "ray[rllib]>=2.43.0",
  "rlgym[all]>=2.0.0",
  "rocketsim>=2.1.1.post1",
  "ruff>=0.11.0",
  "tensorboard>=2.19.0",
  "torch>=2.6.0",
]

[tool.ruff]
line-length = 140
extend-select = ["I"]
